requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aniso8601==7.0.0
attrs==19.1.0
click==7.0
flask-accepts==0.10.0
flask-cors==3.0.8
flask-restful==0.3.7
flask-restplus==0.12.1
flask-restx==0.2.0
flask-script==2.0.6
flask-sqlalchemy==2.4.0
flask==1.1.1
geoalchemy2==0.8.4
grpcio-tools==1.41.0
grpcio==1.41.0
itsdangerous==1.1.0
jinja2==2.11.2
jsonschema==3.0.2
kafka-python==2.0.2
markupsafe==1.1.1
marshmallow-sqlalchemy==0.23.1
marshmallow==3.7.1
protobuf==3.20.1rc1; python_version >= '3.7'
psycopg2-binary==2.8.5
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
setuptools==62.0.0; python_version >= '3.7'
shapely==1.7.0
six==1.15.0
sqlalchemy==1.3.19
werkzeug==0.16.1
